[feat] home layout 변경점 적용
