--- app/src/main/java/com/wyb/wyb_android/data/model/Comfort.kt
@@ -1,7 +1,0 @@
-package com.wyb.wyb_android.data.model
-
-data class Comfort(
-    val range: String,
-    val comfort: String,
-    val innerList: MutableList<Challenge>
-)
