+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/challengeFragment"
-        android:name="com.wyb.wyb_android.ui.home.ChallengeFragment"
+        android:name="com.wyb.wyb_android.ui.home.HomeFragment"
         android:label="ChallengeFragment"
-        tools:layout ="@layout/fragment_challenge"/>
+        tools:layout ="@layout/fragment_home"/>
 </navigation>
